Creando controlador torre
